refactor: extract computeDiffs stage from main.ts
Move the diff-computation pipeline into a side-effect-free computeDiffs() so the e2e test can drive it without a PR context. run() now composes computeDiffs() with postDiffComment(); behavior is unchanged.
